(refactor) product and productSupplier validations
