+++ src/main/java/com/utn/ProgIII/dto/ProductDTO.java
@@ -1,16 +1,13 @@
 package com.utn.ProgIII.dto;
 
-import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 
 @Builder
 
 public record ProductDTO (Long idProduct,
 
-                          @NotBlank
                           String name,
 
-                          @NotBlank
                           String status) {
