--- src/main/java/com/utn/ProgIII/exceptions/InvalidCharactersException.java
@@ -1,7 +1,0 @@
-package com.utn.ProgIII.exceptions;
-
-public class InvalidCharactersException extends RuntimeException {
-    public InvalidCharactersException(String message) {
-        super(message);
-    }
-}
