+++ src/main/java/com/utn/ProgIII/exceptions/InvalidProductStatusException.java
@@ -1,0 +1,8 @@
+package com.utn.ProgIII.exceptions;
+
+public class InvalidProductStatusException extends RuntimeException {
+
+    public InvalidProductStatusException(String message) {
+        super(message);
+    }
+}
